switch to sprite_recycle #1439
